+++ src/Movim/Bootstrap.php
@@ -109,6 +109,9 @@ class Bootstrap
         define('WIDGETS_PATH', DOCUMENT_ROOT . '/app/widgets/');
         define('MOVIM_SQL_DATE', 'Y-m-d H:i:s');
 
+        define('DEFAULT_PICTURE_FORMAT', 'webp');
+        define('DEFAULT_PICTURE_QUALITY', 95);
+
         define('API_SOCKET', CACHE_PATH . 'socketapi.sock');
     }